requirements.txt
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.1
enum34==1.1.10
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
ipaddress==1.0.23
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
pyOpenSSL==19.1.0
requests==2.24.0
selenium==3.141.0
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
APScheduler==3.0.0
Flask-APScheduler==1.11.0
schedule==0.6.0
psycopg2-binary==2.8.6
beautifulsoup4==4.4.1
requests==2.23.0
